refactor(infrastructure): split TrainerRelationshipRepository into focused partials
Extract dashboard query logic and relationship link methods into dedicated partials. Main repository reduced to core CRUD operations, keeping GetDashboardTraineesAsync as specified.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
